ya casi termino

+++ src/main/java/com/company/bitacora/backend/model/dao/EquipoDao.java
@@ -3,5 +3,5 @@ package com.company.bitacora.backend.model.dao;
 import com.company.bitacora.backend.model.Equipo;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EquipoDao extends CrudRepository<Equipo, Integer> {
+public interface EquipoDao extends CrudRepository<Equipo, Long> {
 }
