+++ src/main/java/com/company/bitacora/backend/model/dao/SalonDao.java
@@ -3,5 +3,5 @@ package com.company.bitacora.backend.model.dao;
 import com.company.bitacora.backend.model.Salon;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SalonDao extends CrudRepository<Salon, Integer> {
+public interface SalonDao extends CrudRepository<Salon, Long> {
 }
