+++ src/main/java/com/company/bitacora/backend/model/dao/SoporteTecnicoDao.java
@@ -3,5 +3,5 @@ package com.company.bitacora.backend.model.dao;
 import com.company.bitacora.backend.model.SoporteTecnico;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SoporteTecnicoDao extends CrudRepository<SoporteTecnico, Integer> {
+public interface SoporteTecnicoDao extends CrudRepository<SoporteTecnico, Long> {
 }
