+++ src/main/java/com/company/bitacora/backend/model/dao/SoporteTecnico_BitacoraDao.java
@@ -3,5 +3,5 @@ package com.company.bitacora.backend.model.dao;
 import com.company.bitacora.backend.model.SoporteTecnico_Bitacora;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SoporteTecnico_BitacoraDao extends CrudRepository<SoporteTecnico_Bitacora, Integer> {
+public interface SoporteTecnico_BitacoraDao extends CrudRepository<SoporteTecnico_Bitacora, Long> {
 }
